Extract SttpOpenAPICirceEncoders to own file
